GUI/REND: Move Vjoy OSD to gui_renderer.cpp, abstract away from rend_if
