+++ libswirl/rend/gl4/gl4rend.cpp
@@ -538,8 +538,6 @@ static bool gl4_create_resources()
 
 	create_modvol_shader();
 
-	gl_load_osd_resources();
-
 	// Create the buffer for Translucent poly params
 	glGenBuffers(1, &gl4.vbo.tr_poly_params);
 	// Bind it
@@ -1040,7 +1038,6 @@ struct gl4rend : Renderer
 
 		CollectCleanup();
 
-		gl_free_osd_resources();
 		free_output_framebuffer();
 		gl41_term();
 	}
@@ -1051,14 +1048,6 @@ struct gl4rend : Renderer
 	bool RenderLastFrame() { return gl4_render_output_framebuffer(); }
 
 	void Present() { os_gl_swap(); }
-
-	void DrawOSD(bool clear_screen)
-	{
-		glBindVertexArray(gl4.vbo.main_vao);
-		glBindBuffer(GL_ARRAY_BUFFER, gl4.vbo.geometry); glCheck();
-
-		OSD_DRAW(clear_screen);
-	}
 
 	virtual u32 GetTexture(TSP tsp, TCW tcw) {
 		return gl_GetTexture(vram, tsp, tcw);
